+++ internal/acceptance_tests/blocks/logging_honeycomb_empty_format.tf
@@ -1,0 +1,10 @@
+resource "fastly_service_logging_honeycomb" "test" {
+  service_id = fastly_service_cdn.test.id
+  version    = {{.SERVICE_VERSION}}
+  name       = "{{.LOGGING_HONEYCOMB_NAME}}"
+  dataset    = "test-dataset"
+  authentication = {
+    token = "test-honeycomb-key"
+  }
+  format = ""
+}