+++ internal/acceptance_tests/blocks/logging_honeycomb_nested_updated.tf
@@ -1,0 +1,9 @@
+logging_honeycomb {
+  name    = "{{.LOGGING_HONEYCOMB_NAME}}"
+  dataset = "updated-dataset"
+  authentication = {
+    token = "updated-honeycomb-key"
+  }
+  processing_region = "eu"
+  format_version     = 2
+}