+++ internal/provider/provider.go
@@ -100,6 +100,7 @@ import (
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/logginggooglepubsub"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/logginggrafanacloudlogs"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/loggingheroku"
+	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/logginghoneycomb"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/logginghttps"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/loggingkafka"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/loggingloggly"
@@ -251,6 +252,7 @@ func (p *fastlyProvider) Resources(_ context.Context) []func() resource.Resource
 		logginggooglepubsub.NewResource,
 		logginggrafanacloudlogs.NewResource,
 		loggingheroku.NewResource,
+		logginghoneycomb.NewResource,
 		logginghttps.NewResource,
 		loggingkafka.NewResource,
 		loggingloggly.NewResource,
@@ -398,6 +400,7 @@ func (p *fastlyProvider) ListResources(_ context.Context) []func() list.ListReso
 		logginggooglepubsub.NewListResource,
 		logginggrafanacloudlogs.NewListResource,
 		loggingheroku.NewListResource,
+		logginghoneycomb.NewListResource,
 		logginghttps.NewListResource,
 		loggingkafka.NewListResource,
 		loggingloggly.NewListResource,